requirements.txt
SQLAlchemy-Utils~=0.41.2
SQLAlchemy~=2.0.39
python-dotenv~=1.0.1
Flask~=3.1.0
psycopg2-binary~=2.9
Werkzeug~=3.1.3
WTForms~=3.2.1
Flask-Login~=0.6.3